requirements.txt
streamlit
psycopg2-binary
pandas
plotly
altair
scikit-learn
python-dotenv